Renaming the theme file
